requirements.txt
# --- UI ---
streamlit>=1.20

# --- Core AI / LangChain ecosystem ---
langchain>=0.2.0
langchain-core>=0.2.0
langchain-community>=0.2.0
langchain-openai>=0.1.0
openai>=1.0.0

# --- Google Sheets integration ---
gspread>=5.8.0
google-auth>=2.20.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0

# --- Utilities ---
pandas>=2.0
python-dotenv>=1.0
requests>=2.31.0